Add file api-example.sh
Converted and expanded the api-example.py into a bash script that takes preset (optional) and prompt as arguments. More than just an example, this can be used in shell scripts and loops to automate text generation.

Simple example:

preset='NovelAI-Sphinx Moth'
prompt=$'This is a conversation with your Assistant. The Assistant is very helpful and is eager to chat with you and answer your questions.\nYou: Explain quantum computing in simple terms\nAssistant:'
./api-example.sh "$preset" "$prompt"