Add check to skip lines starting with '#'
Add check to skip lines from the file that start with `#`

In the function `get_args_from_file()`, added a check to skip any lines that start with the `#` character. This change was made to exclude comments from the list of arguments that are extracted from a file. The new check uses the `startswith()` method to check if the first character of the line is `#`, and if it is, the line is skipped and the loop moves on to the next line. This improves the functionality of the function by ensuring that comments in the file are not treated as arguments.