Refactor arg parsing & added config opts
This commit refactors the way command-line arguments and configuration files are loaded. 

Previously, the code used a combination of arguments from `webui.conf` and the command line. The refactored code now separates configuration file loading and argument parsing into two distinct steps, making it easier to manage both.

The changes involve moving the `parser.parse_args()` function to the beginning of the code and creating a new `combined_args` list by parsing the configuration files and command-line arguments separately. If the `--no_config` option is not used and the `webui.conf` configuration file exists, it is added to the list of configuration files. Any additional configuration files passed using the `--config` option are also added to the list. If a configuration file is missing, an error message is printed, and the application exits with an error code.

Once the configuration files are parsed, the command-line arguments are added to the `combined_args` list. Finally, the `parser.parse_args()` function is called with the `combined_args` list to parse all the arguments.

These changes simplify the code, making it easier to manage the arguments and configuration files, and reducing the potential for errors.