Updated Script.py to add a dropdown
Implemented a new feature to the script that incorporates a checkbox and a dropdown to enable users to switch between predefined biases in the dropdown or input custom bias via a textbox. Upon execution, the code generates a new "bias_options.txt" file in the extensions directory, which stores the option values for the dropdown. This enables users to modify the file and add their frequently used biases for quick access. Please note that restarting the webUI is necessary for changes made to the txt file to reflect in the dropdown.