script.py
Made a separate button to apply the model it saves time not loading the same model each time you regenerate. I will put all these persistent settings in the same compartment within the row.
I don't know if I need to update something, but the VRAM management can't load and unload the checkpoints for me. Also, the Resize preview is not working :(