Changed prompt processing and added suffixes
Added suffixes to the SD side prompts, which can be a toggle for specific baked-in tags in the nsfw_prompts and anti_nsfw_prompts parameters. Added character self detection in prompt. On self detection, prompt change to ask character to describe its clothing and a marker is set to add positive_sd tag in character json to positive SD prompt and negative_sd tag in character json to negative SD prompt.  If SD translations is toggled, extension will look into extensions/sd_api/pictures/translations.json and will add tags to SD side prompt if a string in the descriptive_word array matches a string in either the prompt sent to the text generation AI or the response from the text generation AI. This is useful, for instance, if you want the extension to recognize the word "tennis" to trigger a tennis focussed LORA in SD and add tags you would always want to be in a tennis-related image.