Update README.MD
Adjusted the sample for the translations, to follow the suggestions from @ClayShoaf , and to add a LORA to the example. I prefer adding the LORA example to the readme than the translations.json file as it could cause issues if applied by users who don't have that LORA, or by users who don't edit the translations.json file.