It finally works!!! WOOHOO!!
After a full day's worth of work, I actually have something functional.
It's not pretty and it's not perfect, and it may only have options for
prompts and presets, but it functions as expected. I'll
put more information in a future commit and I need to put some more
comments in, but it works!