Update Script
I added dropdown menus to select all the models, samplers and up-scalers add Hi-res fix, clip-skip, batch size, count, and keywords "SD prompt" for interactive and picture book mode, it also saves the generation parameters on a text file and in stealth mode into the PNG file, for the last features needs this script on the script folder on A1111 and save text chunks on the png to work then you can drag the image into the image info tab and you get the parameters. it works with the dynamic address too.