Add example Guanaco prompts. These are from the training dataset because I couldn't find examples.
Because they are from the training dataset, they do not represent actual model performance. But they do perfectly represent model prompt formatting.

Guanaco is super weird in how it expects prompts to be formatted. For a question with no chat history, it is like Alpaca but without the context line at the top. The question can also have an optional input. But for a question with chat history, the instruction section becomes the chat history (usually, but not always, with no line breaks between User and Assistant) and the input section becomes with question but with "User:" in front, and there can be an optional "System:" message at the start of the input section which for things like web search results or chat context. And there is a linebreak between the system and user message. And for Asian languages there is no space between the "User:" and the user input, but for European languages there is a space.