Openai ext. compat with lowercase openai role requests
Alpaca format compatibility with openai's 'system', 'assistant' and 'user' roles lowercase.

Alpaca prompt syntax is as follows:
```
Task description...

### Instruction:
Summarize following text.

### Input:
Text to be summarized

### Response:
```

https://github.com/tatsu-lab/stanford_alpaca/blob/65512697dc67779a6e53c267488aba0ec4d7c02a/train.py#L31

https://github.com/ggerganov/llama.cpp/issues/302

Notice how the "roles" (Instruction, Input, Response) have a capital letter.

Same goes when using the following format:
```
Below is an instruction that describes a task. Write a response that appropriately completes the request.

### Instruction:
Write a poem about the transformers Python library. 

### Response:
```

Or for openai's compatibility:
```
Below is an instruction that describes a task. Write a response that appropriately completes the request.

### User:
Write a poem about the transformers Python library. 

### Assistant:
```