Allow extensions to wrap `generate_reply`
**What**

I added a `wrap_method` decorator to `modules/extensions.py`, which
can be used to easily mark methods to be "wrappable" by extensions.

For now, I only marked `generate_reply`, which allows extensions to
modify either the input or the full output of this method.

**Why**

I'm using this app to host models within my team and we currently
monkey-patch utilities into it, which obviously breaks with each update.
The current extension methods don't fulfill our needs and I I also can't
really upstream our changes, because they don't really belong in the
core of this server.

If this is merged, I could also possibly share the extension we use to
log all prompts+results into MLFlow