fix for issue #2475: Streaming api deadlock
The streaming API runs on its own thread.  This single thread can
handle multiple simultaneous requests from different websockets using
asyncio.

Currently, when multiple requests are received at once, we try to
acquire the shared lock shared.generation_lock multiple times in
succession in the same thread, which will cause a deadlock.

To avoid this, we create a decorator which will grab a thread-local
asyncio.Lock before we try to acquire the shared lock.  This will
prevent the deadlock.

This is a little more complicated than it needs to be because
currently there's only one thread, so we could just use a single,
global asyncio lock instead of one in TLS.  However this approach
will scale in the future if we ever want to have multiple threads
handling requests in parallel.