Add perplexity_colors extension
An extension that colors outputs by perplexity and/or probability per token, and computes average perplexity of the message. This is an example of a logits processor extension (though not one that actually modifies the logits). Note that this is designed for dark mode and also doesn't show up correctly when the output includes code blocks. I was also testing out a hover table of top probabilities but couldn't quite get the CSS working, so I excluded it, though I might try again later.