openai: instruction-templates: format system messages inside context template
Adds capability to use `<|system-message|>`  inside the `context` template setting. If present, it will be replaced by the current set of system messages. If present, but no system messages are set, falls back to the new `system_message_default` template setting.

If `<|system-message|>` is not present inside the `context` template setting, the code behaves exactly as it did before.