hugginfce loader disable_exllama
We had discussion about it, but I think that in all use cases, we should simply 'disable_exllama' : True in Transformers. 

It's supposed to be applicable only to quantized 4-bit and not unquantized but loaded as 4-bit using bnb but you know what if I:
- train lora on HF loaded as 4-bit bnb with disable_exllama' : False
and
- interference HF loaded as 4-bit in bnb with that above LORA applied and disable_exllama' : False

I get a really bad results. 

If I disable that with
disable_exllama' : True
and do the same
all is good as it was. 
So call me paranoid.

In worse case disable_exllama' : True will have no effect at all. So I suggest to simply sneak this in - we have exllama loaders anyhow. I don't think we need Transformer loader that "may or may not" use exllama.