Update one_click.py to initialize site_packages_path variable
Added site_packages_path = None

The reason for this is on new installs this variable won't be initialized by the time we hit the if block trying to access the variable. With this cod if this is a brand new install the code will fail the if check instead of throwing an exception, and if the project has already been installed it will pass the if check appropriately.