Text Generation: Abort if EOS token is reached
Currently, the EOS token is sent via a reponse through the API. However,
an EOS token signals the end of streaming or generation. Therefore,
there's no necessity to include the EOS token in an API response and just
cut it off since that's what the EOS token is doing anyways.

Signed-off-by: kingbri <bdashore3@proton.me>