Update script.py
changed the timesleep wating for subprocess to start up from 20 to 25, as it looks like things take a little longer when you load up with a LLM model loaded. This should just avoid 1x warning message appearing that the subprocess hasnt yet started up (which it always does a second after the warning message appears.