New versions TTS released. Updating requirements
They have made 2x updates to the TTS engine in the last day or so https://github.com/coqui-ai/TTS/discussions/3340 and https://github.com/coqui-ai/TTS/discussions/3354

One of them has improvements for the text splitting within the engine. So thought it a good idea to update.

Also improved the checking within the modelsdownload.py to read in the settings from the requirements.txt file and compare against that.