Support running as a git submodule.
Allow `text-generation-webui` to run as a git submodule within a parent git repo.

In standalone git repos, `.git` is a _directory_; in submodules, `.git` is a _text file_ containing a pointer to the commit SHA.

This PR simply checks whether `.git` exists instead of requiring it to be a directory, and explicitly specifies the remote and branch names when pulling.