Update start_windows.bat
Fixed the broken checksum bug where it will always, at all times, inevitably, fail the checksum, because there is either something funky going on with the length of the filename, or it doesn't play well with there being multiple instances on a single machine. 
This has been broken for far too long, so better to just remove it than to leave it broken honestly.