Update ui_model_menu.py blocking the --multi-user access in backend
In the --multi-user mode, the app is blocking the access by disabling the UI components. But with a simple inspect an abuser could enable it and run the action. So we should block the access also in the backend. A simple way is to block it in the event_handlers.
Here specifically the abuser can't unload the model by doing a simple inspect!