Update ui_session.py blocking the --multi-user access in backend
In the --multi-user mode, the app is blocking the access by disabling the UI components. But with a simple inspect an abuser could enable it and run the action. So we should block the access also in the backend.
Here in an abuse scenario the abuser can change the Boolean command-line flags (consider the multi_user checkbox) and then enable the "Apply flags/extensions and restart" button only with a simple web inspect. At the end the app will be restarted in a not multi-user (single user) mode and the abuser now has the full control of the app.